refactor: split team resume facade
