+++ src/features/team-mode/team-state-store/resume-report.ts
@@ -1,0 +1,7 @@
+export interface ResumeReport {
+  resumed: number
+  marked_failed: number
+  marked_orphaned: number
+  cleaned: number
+  errors: Error[]
+}
